refactor: Refactor Persona module
