+++ messenger/persona/service.py
@@ -1,16 +1,4 @@
-from typing import TypedDict
-
-from messenger.repository import unversioned_client
-
-
-class Persona(TypedDict):
-    name: str
-    profile_picture_url: str
-
-
-def create_persona(persona: Persona):
-    r = unversioned_client.post("me/personas", json=persona)
-    return r.json()
+from messenger.persona.repository import Persona, create_persona
 
 
 def create_gau_vuanem_persona():
